# crates/api/openapi.json
{
  "components": {
    "schemas": {
...
      "NoteBacklink": {
        "properties": {
          "snippet": {
            "type": "string"
          },
          "ticketKey": {
            "type": "string"
          }
        },
        "required": [
          "ticketKey",
          "snippet"
        ],
        "type": "object"
      },
      "NoteDetail": {
        "properties": {
          "backlinks": {
            "items": {
              "$ref": "#/components/schemas/NoteBacklink"
            },
            "type": "array"
          },
          "body": {
            "type": "string"
          },
          "reminders": {
            "items": {
              "$ref": "#/components/schemas/Reminder"
            },
            "type": "array"
          },
          "ticketKey": {
            "type": "string"
          },
          "updatedAt": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "ticketKey",
          "body",
          "backlinks",
          "reminders"
        ],
        "type": "object"
      },
...
      "PutNoteBody": {
        "properties": {
          "body": {
            "type": "string"
          },
          "ticketKey": {
            "type": "string"
          }
        },
        "required": [
          "ticketKey",
          "body"
        ],
        "type": "object"
      },
...
      "Reminder": {
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "dueAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "snoozedUntil": {
            "type": [
              "string",
              "null"
            ]
          },
          "state": {
            "description": "`pending` | `done`",
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "ticketKey": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "ticketKey",
          "text",
          "dueAt",
          "state",
          "createdAt"
        ],
        "type": "object"
      },
      "ReminderRefBody": {
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
      "ReminderSnoozeBody": {
        "properties": {
          "id": {
            "type": "string"
          },
          "until": {
            "description": "ISO 8601 instant.",
            "type": "string"
          }
        },
        "required": [
          "id",
          "until"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Workflow",
      "url": "https://github.com/"
    },
...
    "license": {
      "name": "Proprietary"
    },
    "title": "Workflow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me/notes": {
      "get": {
        "operationId": "getNote",
        "parameters": [
          {
            "description": "Jira ticket key",
            "in": "query",
            "name": "ticketKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDetail"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/notes?ticketKey= — the ticket's note (empty body when none).",
        "tags": [
          "notes"
        ]
      },
      "put": {
        "operationId": "putNote",
        "parameters": [
          {
            "description": "IANA timezone for ⏰ parsing",
            "in": "query",
            "name": "tz",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutNoteBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteDetail"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "PUT /me/notes?tz= — upsert the note and re-sync derived reminders/links.",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/me/reminders": {
      "get": {
        "operationId": "listReminders",
        "parameters": [
          {
            "description": "pending | done",
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Reminder"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "GET /me/reminders?state= — the user's reminders, due-date ascending.",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/me/reminders/done": {
      "post": {
        "operationId": "reminderDone",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderRefBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reminder"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "POST /me/reminders/done — mark a reminder done.",
        "tags": [
          "notes"
        ]
      }
    },
    "/api/me/reminders/snooze": {
      "post": {
        "operationId": "reminderSnooze",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderSnoozeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reminder"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "POST /me/reminders/snooze — suppress a reminder until the given instant.",
        "tags": [
          "notes"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Same-origin (the web app proxies /api to this server)",
      "url": "/"
    }
  ],
...
}